+++ Assets/Examples/Scripts/SampleBehaviour4.cs
@@ -34,7 +34,7 @@ public class SampleBehaviour4 : MonoBehaviour
 
     [Label("Button", skinStyle: SkinStyle.Box)]
 
-    [EditorButton(nameof(TestMethod), Tooltip = "Custom Tooltip", ValidateMethodName = nameof(ValidationMethod))]
+    [EditorButton(nameof(TestMethod), Tooltip = "Custom Tooltip", ValidateMethodName = nameof(ValidationMethod), PositionType = ButtonPositionType.Above)]
     [EditorButton(nameof(TestCoroutine), "<b>Test Coroutine</b>", activityType: ButtonActivityType.OnPlayMode)]
     [EditorButton(nameof(TestStaticMethod), activityType: ButtonActivityType.OnEditMode)]
     public int var1;